+++ src/implementation/Client/DaprClient.ts
@@ -6,6 +6,7 @@ import IClientSecret from '../../interfaces/Client/IClientSecret';
 import IClientHealth from '../../interfaces/Client/IClientHealth';
 import IClientMetadata from '../../interfaces/Client/IClientMetadata';
 import IClientSidecar from '../../interfaces/Client/IClientSidecar';
+import IClientActorBuilder from '../../interfaces/Client/IClientActorBuilder';
 import IClient from '../../interfaces/Client/IClient';
 
 import GRPCClientBinding from './GRPCClient/binding';
@@ -16,6 +17,7 @@ import GRPCClientSecret from './GRPCClient/secret';
 import GRPCClientHealth from './GRPCClient/health';
 import GRPCClientMetadata from './GRPCClient/metadata';
 import GRPCClientSidecar from './GRPCClient/sidecar';
+import GRPCClientActor from './GRPCClient/actor';
 import GRPCClient from './GRPCClient/GRPCClient';
 
 import HTTPClientBinding from './HTTPClient/binding';
@@ -26,6 +28,7 @@ import HTTPClientSecret from './HTTPClient/secret';
 import HTTPClientHealth from './HTTPClient/health';
 import HTTPClientMetadata from './HTTPClient/metadata';
 import HTTPClientSidecar from './HTTPClient/sidecar';
+import HTTPClientActor from './HTTPClient/actor';
 import HTTPClient from './HTTPClient/HTTPClient';
 
 import CommunicationProtocolEnum from '../../enum/CommunicationProtocol.enum';
@@ -46,6 +49,7 @@ export default class DaprClient {
   readonly health: IClientHealth;
   readonly metadata: IClientMetadata;
   readonly sidecar: IClientSidecar;
+  readonly actor: IClientActorBuilder;
 
   constructor(
     daprHost: string
@@ -79,6 +83,7 @@ export default class DaprClient {
         this.health = new GRPCClientHealth(client);
         this.metadata = new GRPCClientMetadata(client);
         this.sidecar = new GRPCClientSidecar(client);
+        this.actor = new GRPCClientActor(client); // we use a abstractor here since we interface through a builder with the Actor Runtime
         break;
       }
       case CommunicationProtocolEnum.HTTP:
@@ -94,6 +99,7 @@ export default class DaprClient {
         this.health = new HTTPClientHealth(client);
         this.metadata = new HTTPClientMetadata(client);
         this.sidecar = new HTTPClientSidecar(client);
+        this.actor = new HTTPClientActor(client); // we use a abstractor here since we interface through a builder with the Actor Runtime
         break;
       }
     }